created helper function to get txn resBody
